+++ Example/SugarRecord/Source/Main/ViewController.swift
@@ -6,11 +6,11 @@ class ViewController: UIViewController, UITableViewDataSource, UITableViewDelega
     // MARK: - Attributes
     
     lazy var tableView: UITableView = {
-        let _tableView = UITableView(frame: CGRectZero, style: UITableViewStyle.Plain)
+        let _tableView = UITableView(frame: CGRect.zero, style: UITableViewStyle.plain)
         _tableView.translatesAutoresizingMaskIntoConstraints = false
         _tableView.delegate = self
         _tableView.dataSource = self
-        _tableView.registerClass(UITableViewCell.classForCoder(), forCellReuseIdentifier: "default-cell")
+        _tableView.register(UITableViewCell.classForCoder(), forCellReuseIdentifier: "default-cell")
         return _tableView
     }()
     
@@ -36,18 +36,18 @@ class ViewController: UIViewController, UITableViewDataSource, UITableViewDelega
     
     // MARK: - Setup
     
-    private func setup() {
+    fileprivate func setup() {
         setupView()
         setupTableView()
     }
 
-    private func setupView() {
-        self.view.backgroundColor = UIColor.whiteColor()
+    fileprivate func setupView() {
+        self.view.backgroundColor = UIColor.white
     }
     
-    private func setupTableView() {
+    fileprivate func setupTableView() {
         self.view.addSubview(tableView)
-        self.tableView.snp_makeConstraints { (make) -> Void in
+        self.tableView.snp.makeConstraints { (make) -> Void in
             make.edges.equalTo(self.view)
         }
     }
@@ -55,39 +55,35 @@ class ViewController: UIViewController, UITableViewDataSource, UITableViewDelega
     
     // MARK: - UITableViewDataSource / UITableViewDelegate
     
-    func tableView(tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return 4
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return 3
     }
     
-    func tableView(tableView: UITableView, cellForRowAtIndexPath indexPath: NSIndexPath) -> UITableViewCell {
-        let cell: UITableViewCell = tableView.dequeueReusableCellWithIdentifier("default-cell")!
-        switch indexPath.row {
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        let cell: UITableViewCell = tableView.dequeueReusableCell(withIdentifier: "default-cell")!
+        switch (indexPath as NSIndexPath).row {
         case 0:
             cell.textLabel?.text = "CoreData Basic"
         case 1:
             cell.textLabel?.text = "Realm Basic"
         case 2:
-            cell.textLabel?.text = "CoreData Observable"
-        case 3:
-            cell.textLabel?.text = "Realm Observable"
+            cell.textLabel?.text = "News from API"
         default:
             cell.textLabel?.text = ""
         }
-        cell.accessoryType = UITableViewCellAccessoryType.DisclosureIndicator
+        cell.accessoryType = UITableViewCellAccessoryType.disclosureIndicator
         return cell
     }
     
-    func tableView(tableView: UITableView, didSelectRowAtIndexPath indexPath: NSIndexPath) {
-        tableView.deselectRowAtIndexPath(indexPath, animated: true)
-        switch indexPath.row {
+    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
+        tableView.deselectRow(at: indexPath, animated: true)
+        switch (indexPath as NSIndexPath).row {
         case 0:
             self.navigationController?.pushViewController(CoreDataBasicView(), animated: true)
         case 1:
             self.navigationController?.pushViewController(RealmBasicView(), animated: true)
         case 2:
-            self.navigationController?.pushViewController(CoreDataObservableView(), animated: true)
-        case 3:
-            self.navigationController?.pushViewController(RealmObservableView(), animated: true)
+            self.navigationController?.pushViewController(NewsView(), animated: true)
         default:
             break
         }