+++ SugarRecord/Source/CoreData/Entities/CoreDataObservable.swift
@@ -1,22 +1,22 @@
-#if os(iOS) || os(tvOS) || os(watchOS)
 import Foundation
 import CoreData
+#if os(iOS) || os(tvOS) || os(watchOS)
 
-public class CoreDataObservable<T: NSManagedObject where T:Equatable>: RequestObservable<T>, NSFetchedResultsControllerDelegate {
+public class CoreDataObservable<T: NSManagedObject>: RequestObservable<T>, NSFetchedResultsControllerDelegate where T:Equatable {
 
     // MARK: - Attributes
 
-    internal let fetchRequest: NSFetchRequest
-    internal var observer: (ObservableChange<T> -> Void)?
-    internal let fetchedResultsController: NSFetchedResultsController
+    internal let fetchRequest: NSFetchRequest<NSFetchRequestResult>
+    internal var observer: ((ObservableChange<T>) -> Void)?
+    internal let fetchedResultsController: NSFetchedResultsController<NSFetchRequestResult>
     private var batchChanges: [CoreDataChange<T>] = []
 
 
     // MARK: - Init
 
-    internal init(request: Request<T>, context: NSManagedObjectContext) {
+    internal init(request: FetchRequest<T>, context: NSManagedObjectContext) {
 
-        let fetchRequest: NSFetchRequest = NSFetchRequest(entityName: T.entityName)
+        let fetchRequest: NSFetchRequest = NSFetchRequest<NSFetchRequestResult>(entityName: T.entityName)
         if let predicate = request.predicate {
             fetchRequest.predicate = predicate
         }
@@ -33,10 +33,10 @@ public class CoreDataObservable<T: NSManagedObject where T:Equatable>: RequestOb
 
     // MARK: - Observable
 
-    public override func observe(closure: ObservableChange<T> -> Void) {
+    public override func observe(_ closure: @escaping (ObservableChange<T>) -> Void) {
         assert(self.observer == nil, "Observable can be observed only once")
-        let initial = try! self.fetchedResultsController.managedObjectContext.executeFetchRequest(self.fetchRequest) as! [T]
-        closure(ObservableChange.Initial(initial))
+        let initial = try! self.fetchedResultsController.managedObjectContext.fetch(self.fetchRequest) as! [T]
+        closure(ObservableChange.initial(initial))
         self.observer = closure
         _ = try? self.fetchedResultsController.performFetch()
     }
@@ -50,40 +50,28 @@ public class CoreDataObservable<T: NSManagedObject where T:Equatable>: RequestOb
 
 
     // MARK: - NSFetchedResultsControllerDelegate
-
-    public func controller(controller: NSFetchedResultsController, didChangeObject anObject: AnyObject, atIndexPath indexPath: NSIndexPath?, forChangeType type: NSFetchedResultsChangeType, newIndexPath: NSIndexPath?) {
+    
+    public func controller(_ controller: NSFetchedResultsController<NSFetchRequestResult>, didChange anObject: Any, at indexPath: IndexPath?, for type: NSFetchedResultsChangeType, newIndexPath: IndexPath?) {
         switch type {
-        case .Delete:
-            #if os(watchOS)
-                self.batchChanges.append(.Delete(indexPath!.indexAtPosition(0), anObject as! T))
-            #else
-                self.batchChanges.append(.Delete(indexPath!.row, anObject as! T))
-            #endif
-        case .Insert:
-            #if os(watchOS)
-                self.batchChanges.append(.Insert(newIndexPath!.indexAtPosition(0), anObject as! T))
-            #else
-                self.batchChanges.append(.Insert(newIndexPath!.row, anObject as! T))
-            #endif
-        case .Update:
-            #if os(watchOS)
-                self.batchChanges.append(.Update(indexPath!.indexAtPosition(0), anObject as! T))
-            #else
-                self.batchChanges.append(.Update(indexPath!.row, anObject as! T))
-            #endif
+        case .delete:
+            self.batchChanges.append(.delete(indexPath![0], anObject as! T))
+        case .insert:
+            self.batchChanges.append(.insert(newIndexPath![0], anObject as! T))
+        case .update:
+            self.batchChanges.append(.update(indexPath![0], anObject as! T))
         default: break
         }
     }
 
-    public func controllerWillChangeContent(controller: NSFetchedResultsController) {
+    public func controllerWillChangeContent(_ controller: NSFetchedResultsController<NSFetchRequestResult>) {
         self.batchChanges = []
     }
 
-    public func controllerDidChangeContent(controller: NSFetchedResultsController) {
-        let deleted = self.batchChanges.filter { $0.isDeletion() }.map { $0.index() }
-        let inserted = self.batchChanges.filter { $0.isInsertion() }.map { (index: $0.index(), element: $0.object()) }
-        let updated = self.batchChanges.filter { $0.isUpdate() }.map { (index: $0.index(), element: $0.object()) }
-        self.observer?(ObservableChange.Update(deletions: deleted, insertions: inserted, modifications: updated))
+    public func controllerDidChangeContent(_ controller: NSFetchedResultsController<NSFetchRequestResult>) {
+        let deleted = self.batchChanges.filter { $0.isDeletion }.map { $0.index() }
+        let inserted = self.batchChanges.filter { $0.isInsertion }.map { (index: $0.index(), element: $0.object()) }
+        let updated = self.batchChanges.filter { $0.isUpdate }.map { (index: $0.index(), element: $0.object()) }
+        self.observer?(ObservableChange.update(deletions: deleted, insertions: inserted, modifications: updated))
     }
 
 }