+++ SugarRecord/Source/Foundation/Entities/RequestObservable.swift
@@ -1,28 +1,28 @@
 import Foundation
 
 public enum ObservableChange<T> {
-    case Initial([T])
-    case Update(deletions: [Int], insertions: [(index: Int, element: T)], modifications: [(index: Int, element: T)])
-    case Error(NSError)
+    case initial([T])
+    case update(deletions: [Int], insertions: [(index: Int, element: T)], modifications: [(index: Int, element: T)])
+    case error(Error)
 }
 
-public class RequestObservable<T: Entity>: NSObject {
+open class RequestObservable<T: Entity>: NSObject {
     
     // MARK: - Attributes
     
-    internal let request: Request<T>
+    internal let request: FetchRequest<T>
     
     
     // MARK: - Init
     
-    internal init(request: Request<T>) {
+    internal init(request: FetchRequest<T>) {
         self.request = request
     }
     
     
     // MARK: - Public
     
-    public func observe(closure: ObservableChange<T> -> Void) {
+    open func observe(_ closure: @escaping (ObservableChange<T>) -> Void) {
         assertionFailure("The observe method must be overriden")
     }